Add settings.c and settings.h
