+++ src/cocoa/main.m
@@ -9,6 +9,7 @@
 #include "../ui.h"
 #include "../util.h"
 #include "../utox.h"
+#include "../settings.h"
 
 #include "../av/utox_av.h"
 #include "../ui/dropdowns.h"
