+++ src/friend.c
@@ -10,6 +10,7 @@
 #include "tox.h"
 #include "util.h"
 #include "utox.h"
+#include "settings.h"
 
 #include "av/utox_av.h"
 #include "ui/edits.h"
