+++ src/groups.c
@@ -4,6 +4,7 @@
 #include "logging_native.h"
 #include "main.h"
 #include "util.h"
+#include "settings.h"
 
 #include "ui/edit.h"
 #include "ui/scrollable.h"
