+++ src/logging.c
@@ -1,6 +1,6 @@
 #include "logging_native.h"
 
-#include "main.h"
+#include "settings.h"
 
 int utox_verbosity() {
     return settings.verbose;
