+++ src/main.c
@@ -10,66 +10,9 @@
 #include "theme.h"
 #include "util.h"
 #include "updater.h"
+#include "settings.h"
 
 #include <getopt.h>
-
-SETTINGS settings = {
-    .curr_version = UTOX_VERSION_NUMBER,
-    // .last_version                // included here to match the full struct
-    .show_splash = false,
-
-    // Low level settings (network, profile, portable-mode)
-    .use_proxy      = false,
-    .force_proxy    = false,
-    .enable_udp     = true,
-    .enable_ipv6    = true,
-    .save_encryption = true,
-    .block_friend_requests = false,
-    // .portable_mode               // included here to match the full struct
-
-    .save_encryption    = true,
-
-    .auto_update        = false,
-    .update_to_develop  = false,
-    .send_version       = false,
-
-    .force_proxy        = false,
-    .enable_udp         = true,
-    .enable_ipv6        = true,
-
-    .use_proxy          = false,
-    .proxy_port         = 0,
-
-    // User interface settings
-    .close_to_tray          = false,
-    .logging_enabled        = true,
-    .audiofilter_enabled    = true,
-    .start_in_tray          = false,
-    .start_with_system      = false,
-    .push_to_talk           = false,
-    .audio_preview          = false,
-    .video_preview          = false,
-    .send_typing_status     = false,
-    .use_mini_flist         = false,
-    // .inline_video                // included here to match the full struct
-    .use_long_time_msg      = true,
-    .accept_inline_images   = true,
-
-    // Notifications / Alerts
-    .ringtone_enabled       = true,
-    .status_notifications   = true,
-    .group_notifications    = GNOTIFY_ALWAYS,
-
-    .verbose = 1,
-
-    // .theme                       // included here to match the full struct
-    // OS interface settings
-    .window_height        = MAIN_HEIGHT,
-    .window_width         = MAIN_WIDTH,
-    .window_baseline      = 0,
-
-    .window_maximized     = 0,
-};
 
 /* The utox_ functions contained in src/main.c are wrappers for the platform native_ functions
  * if you need to localize them to a specific platform, move them from here, to each
