+++ src/messages.c
@@ -14,6 +14,7 @@
 #include "tox.h"
 #include "util.h"
 #include "utox.h"
+#include "settings.h"
 
 #include "ui/contextmenu.h"
 #include "ui/draw.h"
