+++ src/tox_callbacks.c
@@ -8,6 +8,7 @@
 #include "tox.h"
 #include "util.h"
 #include "utox.h"
+#include "settings.h"
 
 #include "av/utox_av.h"
