+++ src/ui/dropdowns.c
@@ -10,6 +10,8 @@
 #include "../theme.h"
 #include "../tox.h"
 #include "../util.h"
+#include "../settings.h"
+
 #include "../av/utox_av.h"
 
 // FIXME: Required for UNUSED()
