+++ src/util.h
@@ -6,7 +6,6 @@
 #include <stddef.h>
 
 typedef struct utox_friend FRIEND;
-typedef struct utox_save UTOX_SAVE;
 
 /*todo: sprint_bytes */
 
@@ -102,12 +101,6 @@ void bgrxtoyuv420(uint8_t *plane_y, uint8_t *plane_u, uint8_t *plane_v, uint8_t 
  */
 void scale_rgbx_image(uint8_t *old_rgbx, uint16_t old_width, uint16_t old_height, uint8_t *new_rgbx, uint16_t new_width,
                       uint16_t new_height);
-
-
-// TODO FIXME this needs to me moved out of here so we can drop the ".h" includes above
-UTOX_SAVE *config_load(void);
-void config_save(UTOX_SAVE *save);
-
 
 // TODO FIXME this needs to me moved out of here so we can drop the ".h" includes above
 /*
