+++ src/windows/main.c
@@ -11,6 +11,7 @@
 #include "../tox.h"
 #include "../util.h"
 #include "../utox.h"
+#include "../settings.h"
 
 #include "../av/utox_av.h"
 #include "../ui/dropdowns.h"
