+++ src/xlib/main.c
@@ -10,6 +10,8 @@
 #include "../tox.h"
 #include "../util.h"
 #include "../utox.h"
+#include "../settings.h"
+
 #include "../av/utox_av.h"
 #include "../ui/draw.h"
 #include "../ui/dropdowns.h"
